activity_builder/requirements.txt
# Core dependencies for Athlete Insight Analytics Platform

# ChromaDB for vector database and semantic search
chromadb>=0.4.0

# Data science and analysis libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Environment and configuration
python-dotenv>=0.19.0

# HTTP client for Ollama API calls
requests>=2.28.0

# Web framework for UI (optional)
flask>=3.0.0

# Date and time handling
python-dateutil>=2.8.0

# For CSV and data processing
openpyxl>=3.1.0

# For enhanced data analysis (optional)
scipy>=1.9.0
scikit-learn>=1.1.0

# For command line argument parsing (built-in but explicit)
argparse

# For path handling (built-in but explicit) 
pathlib

# Optional: For advanced NLP features
# sentence-transformers>=2.2.0

# Optional: FastAPI for advanced web services
# fastapi>=0.100.0
# uvicorn>=0.18.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
# uvicorn>=0.20.0
